requirements.txt
graphviz==0.20.1
matplotlib==3.10.5
networkx==3.1
numpy==2.3.2
onnx==1.16.2
torch
torchvision
z3_solver==4.12.2.0
antlr4-python3-runtime==4.9.2
typer